requirements.txt
requests==2.22.0
progress==1.5
pycrypto==2.6.1
m3u8==0.7.1
